Extract Select Chores
